# Optimize the C++ version somewhat