Here is an optimized version of the C++ code. The main reason it was slow was the same as for C: the standard library functions for parsing data are fairly slow.

I also updated the code to optionally use Abseil's fast hash map, as the one in the standard library is unfortunately known to be slow. Abseil is a bit of a pain to use as a dependency so it can't be used in the test directly as it just compiles things with a single `g++` invocation.

The timings for optimized builds on my machine are as follows:

- C: 0.15s
- C++: 0.37s
- C++ with Abseil: 0.29s
- C++ with Abseil and LTO: 0.25s
- C++ with Abseil and Clang: 0.24s

 I did not try Clang + LTO because the self-built Clang trunk I had was wonky for some reason.