@benhoyt 2021-03-15
Nice, thanks for this! I'll take a look and merge (probably later today).